apply: add interactive option
Rather than asking whether or not to apply each patch individually,
list all of the patches at once and allow the user to either apply all
of them, apply none of them, or interactively choose which ones to
apply (a la interactive rebasing).

In addition to choosing which patches to apply, the interactive option
also allows reordering the patches before applying, and choosing to
apply committed or rejected patches.

https://bugzilla.gnome.org/show_bug.cgi?id=657558