Add the user to the cc: on the bug
All the logic to add the user to the cc: list on the bug is done in the
web page for show_bug.cgi and not on the client side, so the option
"Automatically add me to the CC list of bugs I change" in the user's
preferences is not honored. Since that preference defaults to Always
and users will very typically want to see further traffic on their patches,
just always request the user to be added to the Cc.

The main downside to this is that we do another round trip to the server
when creating and attachment, and also potentially create more bug mail.

https://bugzilla.gnome.org/show_bug.cgi?id=663800