Allow setting bz.default-<field name>
If you want to override something like the default version on a per-component
basis, having to set that in the per-tracker configuration is unnatural
and confusing. Allow setting bz.default-version instead. This is interpreted
as overriding the existing configuration of default field values.

Based on a patch from Dan Winship <danw@gnome.org>

https://bugzilla.gnome.org/show_bug.cgi?id=651756