Work around line-break problems
The two lines for the different versions of 'git bz edit' are treated
by asciidoc as a single paragraph and line-wrapped.

In theory, a + at the end of a line is a forced line-break with
asciidoc, but in the manpage generation route, this doesn't work
properly - the magic in a2x that is supposed to turn the asciidoc-br
processing instruction into a line break doesn't work. (We aren't
using a2x at the moment for the conversion, but no real reason to
switch if it doesn't work better.)

So, use a [verse] paragraph instead. This results in mis-indentation,
but slightly better than the previous breakage.

https://bugzilla.gnome.org/show_bug.cgi?id=670686